Moved name generators to their own files to follow single responsibility principle and fixed some typos in ProfileForm component
